% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap}
\alias{dap}
\title{Get DAP}
\usage{
dap(
  URL = NULL,
  catolog = NULL,
  AOI = NULL,
  startDate = NULL,
  endDate = NULL,
  varname = NULL,
  verbose = TRUE
)
}
\arguments{
\item{URL}{local file path or dodC URL}

\item{catolog}{subset of open.dap catolog}

\item{AOI}{sf object}

\item{startDate}{start date (YYYY-MM-DD)}

\item{endDate}{end date (YYYY- MM-DD)}

\item{varname}{name of variable to extract. If NULL, then get all}

\item{verbose}{Should dap_summary be printed?}
}
\value{
data.frame
}
\description{
Define and get data from a DAP resource
}
\details{
Wraps dap_get and dap_crop into one.
If AOI is NULL no spatial crop is executed. If startDate AND endDate are NULL, no temporal crop is executed. If just endDate is NULL it defaults to the startDate.
}
