% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dap_utils.R
\name{dap.summary}
\alias{dap.summary}
\title{Print Summary Information About a OpenDAP Resource}
\usage{
dap.summary(dap)
}
\arguments{
\item{dap}{data.frame from catolog or dap_crop}
}
\description{
Print summary information about a DAP summary
}
