% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dap_utils.R
\name{dap_crop}
\alias{dap_crop}
\title{Crop DAP file}
\usage{
dap_crop(
  URL = NULL,
  catolog = NULL,
  AOI = NULL,
  startDate = NULL,
  endDate = NULL,
  grid_path = "/Users/mjohnson/github/opendap.catalog/cat_grids.json"
)
}
\arguments{
\item{URL}{local file path or dodC URL}

\item{catolog}{subset of open.dap catolog}

\item{AOI}{sf object}

\item{startDate}{start date (YYYY-MM-DD)}

\item{endDate}{end date (YYYY- MM-DD)}

\item{grid_path}{path to grid json}
}
\value{
data.frame
}
\description{
Crop an OpenDAP resource file to a given AOI and time bound
}
\details{
if AOI is NULL no spatial crop is executed. If startDate AND endDate are NULL, no temporal crop is executed. If just endDate is NULL it defaults to the startDate.
}
