% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dap_xyxv}
\alias{dap_xyxv}
\title{Get XYTV data from DAP URL}
\usage{
dap_xyxv(obj)
}
\arguments{
\item{obj}{an OpenDap URL or NetCDF object}
}
\value{
data.frame with (varname, X_name, Y_name, T_name)
}
\description{
Get XYTV data from DAP URL
}
