% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calls.R
\name{.local_reference_extent}
\alias{.local_reference_extent}
\title{Align Grid to DAP Resource}
\usage{
.local_reference_extent(AOI, ref)
}
\arguments{
\item{AOI}{'sf' or SpatVect object}

\item{ref}{catalog entries for data source}
}
\value{

}
\description{
Align Grid to DAP Resource
}
