% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calls.R
\name{.local_time_extent}
\alias{.local_time_extent}
\title{Align Time to DAP Resource}
\usage{
.local_time_extent(startDate, endDate = NULL, ref)
}
\arguments{
\item{startDate}{start date (YYYY-MM-DD)}

\item{ref}{catalog entries for data source}

\item{enDate}{start date (YYYY-MM-DD). If NULL, then only startDate is found.}
}
\value{

}
\description{
Align Time to DAP Resource
}
