% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{.query}
\alias{.query}
\title{Internal query}
\usage{
.query(x, query, subs = NULL)
}
\arguments{
\item{x}{catalog data.frame}

\item{query}{User query}

\item{subs}{subset of x by column (optional)}
}
\value{
data.frame
}
\description{
Internal query
}
