% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.resource_grid}
\alias{.resource_grid}
\title{Extract grid metadata from NC Pointer}
\usage{
.resource_grid(nc, X_name = NULL, Y_name = NULL)
}
\arguments{
\item{nc}{"NetCDF" object which points to the NetCDF dataset. Found with RNetCDF::open.nc.}

\item{X_name}{Name of X diminion. If NULL it is found}

\item{Y_name}{Name of Y diminion. If NULL it is found}
}
\value{
list with (proj, ext, and diminsion)
}
\description{
Extract grid metadata from NC Pointer
}
