% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.resource_time}
\alias{.resource_time}
\title{Extract time metadata from NC Pointer}
\usage{
.resource_time(nc, T_name = NULL)
}
\arguments{
\item{nc}{"NetCDF" object which points to the NetCDF dataset. Found with RNetCDF::open.nc.}

\item{T_name}{Name of T diminion. If NULL it is found}
}
\value{
list with (duration, interval, and nT)
}
\description{
Extract time metadata from NC Pointer
}
