% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grid_meta}
\alias{grid_meta}
\title{Add Grid Metadata}
\usage{
grid_meta(raw)
}
\arguments{
\item{raw}{a data.frame}
}
\value{
data.frame (raw, + diminsion, proj, ext, X_name, Y_name, T_name)
}
\description{
Add Grid Metadata
}
