% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R, R/data_documentation.R
\docType{data}
\name{grids}
\alias{grids}
\title{OpenDAP Catolog Grids}
\format{
An data.frame

An data.frame
}
\source{
Build in package

Build in package
}
\usage{
data(grids)

data(grids)
}
\description{
OpenDAP Catolog Grids

OpenDAP Catolog Grids
}
\keyword{datasets}
