% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.R
\name{materilize_grid}
\alias{materilize_grid}
\title{Materialize Grid from File or inputs}
\usage{
materilize_grid(
  file = NULL,
  ext = NULL,
  diminsion = NULL,
  projection = NULL,
  fillValue = NULL
)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc), or SpatRast object}

\item{ext}{extent (xmin, xmax, ymin, ymax) in some coordinate system}

\item{diminsion}{dimension (number of columns, number of rows)}

\item{projection}{projection - the actual coordinate system}
}
\value{
SpatRaster object
}
\description{
Returns an empty SpatRast obj with prescibed properties
}
