% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R, R/data_documentation.R
\docType{data}
\name{params}
\alias{params}
\title{OpenDAP Catolog Parameters}
\format{
An data.frame

An data.frame
}
\source{
Build in package

Build in package
}
\usage{
data(params)

data(params)
}
\description{
OpenDAP Catolog Parameters

OpenDAP Catolog Parameters
}
\keyword{datasets}
