% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{parse_date}
\alias{parse_date}
\title{Parse Dates from duratoon and interval}
\usage{
parse_date(duration, interval)
}
\arguments{
\item{duration}{time durations}

\item{interval}{time interval}
}
\description{
Parse Dates from duratoon and interval
}
