% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{popRast}
\alias{popRast}
\title{Create and populate SpatRast}
\usage{
popRast(
  vals,
  xmin,
  xmax,
  ymin,
  ymax,
  ncols,
  nrows,
  time,
  crs,
  names = NULL,
  toptobottom = FALSE
)
}
\arguments{
\item{vals}{numeric vector}

\item{xmin}{min X}

\item{xmax}{max X}

\item{ymin}{min Y}

\item{ymax}{max Y}

\item{ncols}{grid columns}

\item{nrows}{grid rows}

\item{time}{time layers}

\item{crs}{CRS}

\item{names}{name of layers}

\item{toptobottom}{read top to bottom?}
}
\value{
SpatRaster
}
\description{
Create and populate SpatRast
}
