% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dap_utils.R
\name{print.dap}
\alias{print.dap}
\title{Print Summary Information About a OpenDAP Resource}
\usage{
\method{print}{dap}(dap)
}
\arguments{
\item{dap}{data.frame from catolog or dap_crop}
}
\value{

}
\description{
Print summary information about a DAP summary
}
