% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dap_file}
\alias{read_dap_file}
\title{Read from a OpenDAP landing page}
\usage{
read_dap_file(URL, id)
}
\arguments{
\item{URL}{URL to OpenDap resource}

\item{id}{character. Uniquely named dataset identifier}
}
\value{
data.frame with (varname, X_name, Y_name, T_name, URL, id), grid, and time
}
\description{
Reads an OpenDap resources and returns metadata
}
