% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_tds}
\alias{read_tds}
\title{Read from a THREDDS catalog HTML page}
\usage{
read_tds(URL, id)
}
\arguments{
\item{URL}{URL to THREDDS catalog}

\item{id}{character. Uniquely named TDS identifier}
}
\value{
data.frame with (link, URL, id)
}
\description{
Scraps a TDS landing page for avalaible datasets
}
