% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Search Catalog}
\usage{
search(query = NULL, source = NULL)
}
\arguments{
\item{query}{a string to search for.}

\item{source}{If the source id you want is known, provide it to expedite search}
}
\value{
data.frame
}
\description{
Search the data catalog to find an ideal source.
}
\details{
query should be passed as a string or list of key word. If a single string is passed, it will be split at spaces.
}
