% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spatial_intersect}
\alias{spatial_intersect}
\title{Catalog Spatial Intersect}
\usage{
spatial_intersect(x, AOI, merge = TRUE)
}
\arguments{
\item{x}{a catalog element}

\item{AOI}{an AOI sf object}

\item{merge}{should the grid meta data be merged with the reduce catalog (default = TRUE)}
}
\value{

}
\description{
Reduce and input catalog to those elements that intersect a given AOI
}
