% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dap_utils.R
\name{var_to_terra}
\alias{var_to_terra}
\title{Var Array to Terra}
\usage{
var_to_terra(var, dap)
}
\arguments{
\item{var}{array}

\item{dap}{dap description}
}
\value{
SpatRast
}
\description{
Var Array to Terra
}
