% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_to_terra2}
\alias{var_to_terra2}
\title{Variable Array to SpatRast 2}
\usage{
var_to_terra2(dap)
}
\arguments{
\item{dap}{dap description}

\item{var}{numeric array}
}
\value{
SpatRast
}
\description{
Variable Array to SpatRast 2
}
