% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variable_meta}
\alias{variable_meta}
\title{Add Variable Metadata}
\usage{
variable_meta(raw, verbose = TRUE)
}
\arguments{
\item{raw}{a data.frame}

\item{verbose}{should messaging be displayed?}
}
\value{
data.frame
}
\description{
Add Variable Metadata
}
