% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{vrt_crop_get}
\alias{vrt_crop_get}
\title{Terra Based Cropping}
\usage{
vrt_crop_get(
  URL = NULL,
  catalog = NULL,
  AOI = NULL,
  g = NULL,
  varname = NULL,
  start = NULL,
  end = NULL,
  toptobottom = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{URL}{a single or set of URLs}

\item{catalog}{a single or set of catalog elements}

\item{AOI}{a spatial boundary}

\item{varname}{a single or set of variables to collect, in NULL all are kept}

\item{start}{for non "dated" items, start can be called by index}

\item{end}{for non "dated" items, end can be called by index}

\item{toptobottom}{should data be inversed?}

\item{verbose}{should warning be emmited?}

\item{grid}{a list containing an extent (), and}
}
\description{
Terra Based Cropping
}
