% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap_crop.R
\name{dap_crop}
\alias{dap_crop}
\title{Crop DAP file}
\usage{
dap_crop(
  URL = NULL,
  catalog = NULL,
  AOI = NULL,
  startDate = NULL,
  endDate = NULL,
  varname = NULL,
  verbose = TRUE
)
}
\arguments{
\item{URL}{local file path or dodC URL}

\item{catalog}{subset of open.dap catalog}

\item{AOI}{sf object}

\item{startDate}{start date (YYYY-MM-DD)}

\item{endDate}{end date (YYYY- MM-DD)}

\item{varname}{name of variable to extract. If NULL, then get all}

\item{verbose}{Should dap_summary be printed?}
}
\value{
data.frame
}
\description{
Crop an OpenDAP resource file to a given AOI and time bound
}
\details{
if AOI is NULL no spatial crop is executed. If startDate AND endDate are NULL, no temporal crop is executed. If just endDate is NULL it defaults to the startDate.
}
