% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dap_utils.R
\name{dap_get}
\alias{dap_get}
\title{Download DAP resource}
\usage{
dap_get(dap, varname = NULL)
}
\arguments{
\item{dap}{data.frame from catolog or dap_crop}

\item{varname}{name of variable to extract. If NULL, then get all}
}
\value{
SpatRaster
}
\description{
Download DAP resource
}
