% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dap_xyxv}
\alias{dap_xyxv}
\title{Retrieve XYTV data from DAP URL}
\usage{
dap_xyxv(url)
}
\arguments{
\item{url}{a OpenDap URL}
}
\value{
data.frame with (varname, X_name, Y_name, T_name)
}
\description{
Retrieve XYTV data from DAP URL
}
