% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{grids}
\alias{grids}
\title{OpenDAP Catolog Grids}
\format{
An data.frame
}
\source{
Build in package
}
\usage{
data(grids)
}
\description{
OpenDAP Catolog Grids
}
\keyword{datasets}
