% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{params}
\alias{params}
\title{OpenDAP Catalog Parameters}
\format{
An data.frame
}
\source{
Build in package
}
\usage{
data(params)
}
\description{
OpenDAP Catalog Parameters
}
\keyword{datasets}
